pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Receipt expense tracker with AI-powered categorization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
    "pytest-playwright>=0.4.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

